### 2019-09-16 #342 @leonawicz
submission: eia
Submitting Author: Matthew Leonawicz (@leonawicz)
Repository:   https://github.com/leonawicz/eia
Version submitted:   0.3.3
Version accepted: TBD

---

```
Package: eia
Title: API Wrapper for 'US Energy Information Administration' Open Data
Version: 0.3.3
    person("Matthew", "Leonawicz", email = "matt_leonawicz@esource.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0001-9452-2771")),
    person(given = "E Source", role = c("cph", "fnd"))
    )
Description: Provides API access to data from the 'US Energy Information Administration' ('EIA') <https://www.eia.gov/>.
    Use of the API requires a free API key obtainable at <https://www.eia.gov/opendata/register.php>.
    The package includes functions for searching 'EIA' data categories and importing time series and geoset time series datasets.
    Datasets returned by these functions are provided in a tidy format or alternatively in more raw form.
    It also offers helper functions for working with 'EIA' date strings and time formats and for inspecting different summaries of series metadata.
    The package also provides control over API key storage and caching of API request results.
    httr,
    memoise,
    lubridate
    covr,
    lintr,
    tidyr
```

The package specifically is for retrieving energy-related datasets from the US Energy Information Administration (EIA) open data API.

Researchers, data analysts, etc. in academia, government and the energy industry and related industries. While it is the US EIA, international energy datasets are available through the API as well.

There are two other overlapping R packages. (1) `EIAdata`, also on CRAN. This package provides basic data access, but data formats and other options are more limited. It does not cover as many API endpoints or offer as much functionality. (2) There is also a different `eia`, github only: https://github.com/krose/eia. It is similar to (1) in its offerings and level of development. Both are fairly minimalist packages that have been around for about five years (around when the EIA API was first launched, I believe) but have not been developed much further since initial releases.

This submission for `eia` (CRAN) is for a package with greater coverage in terms of functionality, API endpoint access, documentation, and robust test coverage and error handling, In comparison to the others it also offers three data output format options for users working with the API data in different contexts (for example, strict R usage or wanting pure JSON results to pipe to other software applications). You can control the level of "tidy"-ness. It also includes a user agent in calls and takes other steps to play well with the API.

Note added by author: This particular API requires users to use their own API key. I cannot run function examples or unit tests on CRAN, but all examples and unit tests run successfully in multiple other environments, on local and remote systems. Full test suite also runs on Travis-CI where I am able to import an encrypted key. Previously for CRAN purposes, a no-key copy of the main vignette was included (API call examples not executed) in the package, while the main and all additional vignettes showing executed API calls and results are part of the pkgdown web documentation. Also for CRAN purposes, function examples in the R package source code are not run if they require an API key to be present.
