### 2019-09-18 #342 @melvidoni
### Editor checks:

---

#### Editor comments

Hello @leonawicz, thanks for submitting to rOpenSci. I'll be your handling editor. The following is the output of `goodpractice::gp()`. Please, take a look at all of this, and let me know once it is done, so I can check again and start looking for reviewers if everything is fixed.

```
── GP eia ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

    R/categories.R:34:NA
    R/categories.R:35:NA
    R/categories.R:36:NA
    R/categories.R:40:NA
    R/categories.R:41:NA
    ... and 227 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be added to the package when you perform `R CMD build`
    on it.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter
    than 80 characters

    R\categories.R:111:1
    R\categories.R:114:1
    R\geoset.R:56:1
    R\key.R:72:1
    R\key.R:85:1
    ... and 3 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R\categories.R:46:18
    R\categories.R:48:19
    R\geoset.R:86:18
    R\series.R:58:10
    R\series.R:84:26
    ... and 3 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error prone and result 1:0 if the expression on the
    right hand side is zero. Use seq_len() or seq_along() instead.

    R\series.R:71:20
    R\series.R:168:17
    R\series.R:218:18

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no redirection of stdout/stderr. Hmm ... looks like a package Error in texi2dvi(file = file, pdf = TRUE,
    clean = clean, quiet = quiet, : pdflatex is not available Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not available
    Error in running tools::texi2pdf() You may want to clean up by 'rm -Rf C:/Users/e95207/AppData/Local/Temp/Rtmp2hHotb/Rd2pdff30c3f634cfa'
```

---

Reviewers: TBD
Due date: TBD
