### 2019-09-18 7df8fe5
Minor code improvements and documentation updates.
