+++ tests/testthat/test-key.R
@@ -5,7 +5,8 @@ test_that("key storage and retrieval works as expected", {
     "Key stored successfully in package environment.",
     "Key stored successfully in options()."
   )
-  wrn <- "EIA API key not found in package environment, global options, or system enivronment variables."
+  wrn <- paste("EIA API key not found in package environment,",
+          "global options, or system enivronment variables.")
 
   expect_warning(x <- eia_get_key("env"), wrn)
   expect_is(x, "NULL")
