+++ NEWS.md
@@ -1,3 +1,8 @@
+# eia 0.3.4
+
+* Minor code improvements.
+* Minor documentation updates.
+
 # eia 0.3.3
 
 * Added a wrapper for the series category endpoint.

[…]
