### 2019-09-18 #342 @leonawicz
Hi @melvidoni Thank you for considering this submission. I have addressed your requested changes, details below, along with a couple questions.

* `Date` field now omitted from `DESCRIPTION`.
* References like `1:nrow` now replaced.
* Lines greater than 80 characters fixed an re-linted.
* Almost all uses of `sapply` switched to `vapply`. There are two cases remaining where this does not work well. An [example](https://stackoverflow.com/questions/42817309/can-vapply-be-used-with-a-variable-length-fun-value) of the basic issue. However, in all cases of `sapply` that occurred in the package, I strictly used it within contexts where it was known to be type-safe in those instances. The two remaining cases require variable length, however. Let me know if you have another suggestion.

### Unit tests

The package has 100% `testthat` unit test coverage, see [here](https://codecov.io/gh/leonawicz/eia?branch=master). The complete set of tests is executed on Travis-CI by including my encrypted API key for Travis [here](https://github.com/leonawicz/eia/blob/master/.travis.yml). This config also shows there are just a few lines in the package that are explicitly excluded from coverage because they are not practically testable.

If the tests are run without an API key present, the vast majority of tests must be skipped. To run/build/check locally an EIA API key must be placed somewhere like your `.Renviron` file as `EIA_KEY=MYKEY`. I could provide you (and reviewers) with my key offline or you could request one from EIA [here](https://www.eia.gov/opendata/register.php) (only requires an email address and they send you a key a moment later).

### R CMD check errors and warnings

I am unable to reproduce these on various Windows or Linux platforms. Is there more information about the environment in which you are building or checking the package? Even without a key present, it will skip the impacted functions and tests and so should still successfully build even in that case.

Thank you! :)
