### 2019-09-18 #342 @melvidoni
Hello @leonawicz, thanks for taking care of this.

Don't worry about the remaining `sapply`, as long as it is tested and controlled. `goodpractices::gp()` now gave me:

```
── GP eia ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R\series.R:84:26
    R\series.R:85:25

──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

So I'm going to start searching for reviewers!
