### 2019-10-18 #342 @daranzolin
Hi all, review is below.

---

Terrific package that is already mostly polished. I learned how to get my API key, peruse the vignettes, search through various child categories, and subset relevant data in less than 30 minutes. I felt free to explore aggressively knowing about the built-in memoization and rate-limiting. Great features. The pattern of searching for parent/child ids and using `eia_series` or `eia_geoset` would soon become familiar and comfortable after more use. The ability to search for regions like "New England" or "USA" was nifty.

Some questions/observation:

* Is there perhaps an easier way to explore/traverse the parent-child relationships? An interactive shiny gadget that cascades select inputs might be cool. Or maybe a `recursive = TRUE` argument in `eia_cats`, similar to `dir` or `list.files`? It might also be useful if I could do something similar to `here::here` with `eia_cats`, e.g. `eia_cats("Electricity", "Total consumption")`.

* You have an example for `unnest`-ing data in the Time series vignette, but not the Geosets vignette. A similar example (since I clicked that one first), would be useful.

* Inconsequential, but there is some inconsistency in the vignettes ending sentences with periods vs. colons.

* I think the README and vignettes would benefit from a visualization. I've used a lot of API wrappers, and seeing what visualizations I can get to helps me understand what data is available quickly.

Some alerts from `goodpractice::gp()`:

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    R\cache.R:5:1
    R\categories.R:5:1
    R\categories.R:6:1
    R\categories.R:10:1
    R\categories.R:11:1

Looks like most pertain to `roxygen` help text and decorators.

And these last two are possibly the result of my machine, not yours, but I'll flag them anyway:

✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.

✖ fix this R CMD check ERROR: Re-running with no redirection of
    stdout/stderr. Hmm ... looks like a package Error in texi2dvi(file = file, pdf =
    TRUE, clean = clean, quiet = quiet, : pdflatex is not available Error in
    texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not
    available Error in running tools::texi2pdf() You may want to clean up by 'rm -Rf
    C:/Users/918831~1/AppData/Local/Temp/RtmpSmuM4O/Rd2pdf934192f4b19'
