### 2019-10-23 #342 @leonawicz
Hi all, I plan to mostly get to work on this once both reviews are in but I can chime in on some initial things.

@daranzolin Thank you for your review. :) Regarding your questions about exploring the category tree, I will continue to give some thought to this but the API is vast and I'd want to be extra careful with recursion, possibly recursing up (`eia_parent_cats` already does this) but not down. I think a Shiny app would be a nice touch but I think users would be much quicker to just browse the existing API where they obtained their key, using the EIA's existing API explorer. Perhaps low return on effort semi-duplicating what it offers.

I like the idea of allowing names like "Electricity" in place of their opaque category ID numbers. However, these numbers are actually better, in part because no one would want to type out many of the category names, which can get really long. Please let me know if you envision something different or I'm misunderstanding. But yes, I agree, the most frustrating part of such a massive collection of datasets behind an API is getting acquainted with the IDs you need for your project. My hunch is most users do this through the API explorer on the website before ever encountering this particular API wrapper package.

I will address the other inconsistencies in the code and documentation you mentioned soon.

I think I fixed all the code lines that were longer than 80 characters, but is it necessary to apply this rule to all the roxygen2 help file text? It has no impact on how help docs render and from an author/maintainer standpoint it's a lot easier to write, read and edit them if the sentences aren't so broken up. I'm happy to do so if required, just let me know.

I agree on the value of visualizations and will be adding some. The console-heavy feel is too much without having any. I'll plan to add some basic visualizations as part of this review process.

Yes I think the errors relate to possibly not having the pdf-creating utility installed on a Windows machine if you are checking/building the package locally.

Thanks so much for your feedback!
