### 2019-10-25 #342 @daranzolin
@leonawicz We recently developed an internal shiny gadget with cascading select filters from user input, and it struck me as a possible application here. Probably superfluous here, though. [The API explorer is sufficient](https://www.eia.gov/opendata/qb.php?category=371) and also shows the category ids (obviating the need to type long strings).
