### 2019-10-25 2a5c8f6
Add basic plots to readme, overview vignette, and series and geoset vignettes. Add initial report function. Update imported and suggested packages. Minor code cleanup, add unit tests.
