+++ man/eia_report.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/reports.R
+\name{eia_report}
+\alias{eia_report}
+\alias{report_drilling_productivity}
+\title{Download data for various EIA reports}
+\usage{
+eia_report(id, ...)
+
+report_drilling_productivity()
+}
+\arguments{
+\item{id}{character, the report ID. See details for the list of available reports.}
+
+\item{...}{arguments passed to individual report data functions.}
+}
+\value{
+a list, typically a list of data frames
+}
+\description{
+These functions download data for various EIA reports found on the EIA website but not necessarily available through the EIA API.
+}
+\details{
+The wrapper function and the individual report functions do not make API calls and do not require an API key.
+}
+\examples{
+\dontrun{
+x <- eia_report("drilling productivity")
+}
+}
