+++ vignettes/geoset.Rmd
@@ -46,6 +46,15 @@ Provide the associated label and `eia_geoset` recognizes subsets of US states. E
 ```{r geoset2}
 (x <- eia_geoset(id, c("AK", "New England"), n = 1))
 x$data[[1]]
+
+region <- "Middle Atlantic"
+x <- eia_geoset(id, region, n = 12)
+select(x, region, data) %>% unnest()
+
+unnest(x, cols = data) %>% 
+  ggplot(aes(date, value, color = region)) +
+  geom_line() + 
+  labs(y = d$units[1], title = "Net electricity generation, all fuels")
 ```
 
 ## Relations
