+++ vignettes/eia.Rmd
@@ -74,10 +74,18 @@ eia_cats(0)
 Time series data is obtained by series ID. Most columns contain metadata. The `data` column contains the time series data.
 
 ```{r series1}
-id <- "ELEC.GEN.ALL-AK-99.A"
-(x1 <- eia_series(id, n = 5))
+library(dplyr)
+library(tidyr)
+library(ggplot2)
 
-x1$data[[1]]
+id <- "ELEC.GEN.ALL-AK-99.A"
+(x <- eia_series(id))
+
+x$data[[1]]
+
+select(x, units, data) %>% unnest(cols = data) %>% 
+  ggplot(aes(date, value)) + geom_line() + 
+  labs(y = d$units[1], title = "Net electricity generation, Alaska, all fuels")
 ```
 
 You can provide arguments like the following:
@@ -102,16 +110,19 @@ This allows you to use the returned results with existing code you may have that
 
 Geosets are metadata structures organizing time series datasets that can be mapped. Arguments to `eia_geoset` are the same as `eia_series` with the addition of `region`. Like `id`, `region` can be a vector. Most of the details are the same as before. 
 
-In the example below using total electricity generation, get the last two data points for each of and two US states. `dplyr` and `tidyr` are used here to clean up the result a bit for purposes of display.
+In the example below using total electricity generation, get the last two data points for each of and two US states. `dplyr` and `tidyr` are used here to clean up the result a bit for purposes of display. `gpplot2` is used to graph the data after it has been unnested for each state.
 
 ```{r geoset1}
-id <- c("ELEC.GEN.ALL-99.A")
+id <- c("ELEC.GEN.ALL-99.M") # monthly
 region <- c("USA-CA", "USA-NY")
-(x <- eia_geoset(id, region, n = 2))
+(x <- eia_geoset(id, region, start = "201801", end = "201812"))
 
-library(dplyr)
-library(tidyr)
 select(x, region, data) %>% unnest()
+
+unnest(x, cols = data) %>% 
+  ggplot(aes(date, value, color = region)) +
+  geom_line() + 
+  labs(y = d$units[1], title = "Net electricity generation, all fuels")
 ```
 
 Another convenience of `eia_geoset` is the ability to provide regions in the following forms:
@@ -125,9 +136,17 @@ These shortcuts make it easier to construct an API call involving several states
 
 ```{r geoset2}
 region <- c("AK", "New England")
-(x <- eia_geoset(id, region, n = 2))
-
+x <- eia_geoset(id, region, n = 2)
 select(x, region, data) %>% unnest()
+
+region <- "Middle Atlantic"
+x <- eia_geoset(id, region, n = 12)
+select(x, region, data) %>% unnest()
+
+unnest(x, cols = data) %>% 
+  ggplot(aes(date, value, color = region)) +
+  geom_line() + 
+  labs(y = d$units[1], title = "Net electricity generation, all fuels")
 ```
 
 Even more convenient is that these names are available in R. See the `datasets::state.*` functions and the geoset vignette.

[…]
