### 2019-10-25 f3fa4e4
Minor vignette cleanup and update docs.
