+++ vignettes/eia-nokey.Rmd
@@ -32,7 +32,7 @@ It is important to store your API key somewhere secure. Do not commit it to a re
 
 You can always provide the `key` argument to every API function call, but you do not have to. There are getter and setter helpers available to make using `eia` functions a more seamless experience.
 
-`eia_set_key` gives you the option of storing your key for the duration of your R session:
+`eia_set_key` gives you the option of storing your key for the duration of your R session.
 
 ```{r key1, eval=FALSE}
 library(eia)
@@ -40,7 +40,7 @@ library(eia)
 # eia_get_key() # retrieve it
 ```
 
-If the key already exists in the system environment and you plan to pass `key` to functions explicitly, you could start as follows:
+If the key already exists in the system environment and you plan to pass `key` to functions explicitly, you could start as follows.
 
 ```{r key2, eval=FALSE}
 key <- Sys.getenv("EIA_KEY")
@@ -57,7 +57,7 @@ Once you have your EIA registered API key and have it in place for your R sessio
 
 It is helpful to be aware of various categories of data that are available. Each category has a unique ID number that is required to access associated data. The first call below does not include an ID. The result is a list of two data frames. The first is metadata associated with that position in the category hierarchy. The second is the child category information.
 
-Here is the top-level category information:
+Here is the top-level category information.
 
 ```{r cats1}
 eia_cats()
@@ -74,10 +74,18 @@ eia_cats(0)
 Time series data is obtained by series ID. Most columns contain metadata. The `data` column contains the time series data.
 
 ```{r series1}
-id <- "ELEC.GEN.ALL-AK-99.A"
-(x1 <- eia_series(id, n = 5))
+library(dplyr)
+library(tidyr)
+library(ggplot2)
 
-x1$data[[1]]
+id <- "ELEC.GEN.ALL-AK-99.A"
+(x <- eia_series(id))
+
+x$data[[1]]
+
+select(x, units, data) %>% unnest(cols = data) %>% 
+  ggplot(aes(date, value)) + geom_line() + 
+  labs(y = x$units[1], title = "Net electricity generation, Alaska, all fuels")
 ```
 
 You can provide arguments like the following:
@@ -102,32 +110,43 @@ This allows you to use the returned results with existing code you may have that
 
 Geosets are metadata structures organizing time series datasets that can be mapped. Arguments to `eia_geoset` are the same as `eia_series` with the addition of `region`. Like `id`, `region` can be a vector. Most of the details are the same as before. 
 
-In the example below using total electricity generation, get the last two data points for each of and two US states. `dplyr` and `tidyr` are used here to clean up the result a bit for purposes of display.
+In the example below using total electricity generation, get the last two data points for each of and two US states. `dplyr` and `tidyr` are used here to clean up the result a bit for purposes of display. `gpplot2` is used to graph the data after it has been unnested for each state.
 
 ```{r geoset1}
-id <- c("ELEC.GEN.ALL-99.A")
+id <- c("ELEC.GEN.ALL-99.M") # monthly
 region <- c("USA-CA", "USA-NY")
-(x <- eia_geoset(id, region, n = 2))
+(x <- eia_geoset(id, region, start = "201801", end = "201812"))
 
-library(dplyr)
-library(tidyr)
 select(x, region, data) %>% unnest()
+
+unnest(x, cols = data) %>% 
+  ggplot(aes(date, value, color = region)) +
+  geom_line() + 
+  labs(y = x$units[1], title = "Net electricity generation, all fuels")
 ```
 
-Another convenience of `eia_geoset` is the ability to provide regions in the following forms:
+Another convenience of `eia_geoset` is the ability to provide regions in the following forms.
 
[…]
