### 2019-10-25 #342 @leonawicz
@ottlngr Sounds good, I will check back next week. Thank you

@melvidoni @daranzolin I have had a chance today to make revisions. The latest version on GitHub cleans up the vignette formatting. I also added basic example ggplot graphs to the readme, getting started vignette, and the most applicable other vignettes- series and geoset, per your suggestion.

Out of general curiosity, can you point me to a resource about this cascading select filter Shiny gadget you mentioned?

Also, after merging your vignette PR, I realized I brought in a new feature with this recent push. I don't want that to confuse anyone, but there is a new function `eia_report`, which is a simple initial version of a basic wrapper function for downloading report data from the EIA website for data from popular canned reports they reelase; data that is not directly available through the API itself. It only offers one report at this time and has unit tests and documentation. This is a very minor feature and I don't want to draw attention to it in the more long-form vignette documentation at this time, but it may grow in the future with more reports as I continue to field user requests and feedback. But I want to mention it just in case you wonder why a new function has appeared.

Thanks! :)
