### 2019-10-27 #342 @ottlngr
---

`eia` is in many ways an exemplary package: Great vignettes and documentation, clean and consistent code as well as state-of-the-art dependecies and 'infrastructure' (tests, CI, `pkgdown`, ...). Using `eia` is pretty straight-forward and I like the way `eia` supplies additional functionality (compared to the raw API) where useful, but generally keeps things simple.

A few questions/remarks:

- I think its a useful feature to basically have three output options (tidy, fromJSON, raw), but I felt a bit uncomfortable using `TRUE`, `FALSE` and `NA` to switch between these choices. For me personally, a setup using `tidy = TRUE|FALSE, raw = FALSE|TRUE` would be more compelling.

- You are using `httr` in the background, but when creating the request url in .eia_url() you rely on `paste0()` instead of using `httr::modify_url()` - is there a reason for this?

- Explicit compliment for setting the user agent globally!

Regarding @daranzolin 's review, I agree with using some visualizations in the vignettes. Also, I wished the exploration of the category tree could be improved - but as you mentioned, @leonawicz, this would result in many additional API requests.

`R CMD CHECKS` does not show any problems on my machine.

Great package, definitely a good supplement to the rOpenSci universe!
