### 2019-10-28 #342 @leonawicz
@melvidoni thank you and @ottlngr thank you for your review, responses below.

### The tidy argument

Here was my thought process when developing this: I agree that `TRUE`, `FALSE`, `NA` may seem a little unusual at first. I also had considered alternatives including `raw = TRUE`. I decided against `raw` anything because it seemed to suggest getting actual `raw`-type data, which some APIs return and that was confusing. I was also looking to avoid creating two separate arguments that in combination determined which of the three output formats the user receives, with one having to be ignored depending on the other.

So I stuck with three values for one argument, and these three values do capture the meaning. I also figure that very few users will ever want something other than `tidy = TRUE` so it is the default and captures the essence of "anything else is *not 'tidy'*." The `NA` option does capture the sense of "tidy or un-tidied R dataset structure just is not applicable here because R is doing nothing to post process your API output; R is just the messenger."

I would like to keep it to a single 3-valued argument, but I am open to suggestions along those lines. I considered things like `output = c("data.frame", "list", "character")` or similar but found this frustrating because it portrays the options as being about nothing more than object class. Or `output = c("tidy", "fromJSON", "string")`, etc. Nothing was feeling quite right. But `tidy = TRUE` seems spot on. Turning that off with `FALSE` seems okay, and `NA` is really only going to matter to hardcore power users who probably don't even want to be accessing the API from R in the first place but just happen to be.

### httr functions

Honestly I didn't even know about `httr::modify_url()`. For package development I not only import as few packages as I can, but when I do, I also import and use few functions. I use `paste`, `paste0` and `sprintf` often in packaged code. I also don't add dependencies like `glue` inside other packages. It wouldn't make a difference here since `httr` is already required. The only distinction between where I relied on `httr` and where I didn't was with the API communications that I didn't otherwise know how to put together. Outside of necessities like that, I tend to use base R for things like string manipulation.

### User agent

Thanks! That was my first time doing that. Just for context, that is one thing that would have to be updated following any repository transfer and it would be up to ROpenSci if the new user agent was the new repo url or something else. I have no opinion on that but it would have to change from what it is now. On that note, the other breaking effect of a transfer would be the encrypted API key in `.travis.yml` would stop working because that is user/repo specific. But this is just details possibly for later.

### Next steps

@melvidoni @daranzolin @ottlngr Okay, I think I have addressed all reviewers' and editor's comments and questions, provided clarifications, etc. I've made updates that you all called for such as the addition of graphs in the README and vignettes, fixes to documentation, etc.

There wasn't a lot of explicit "change this, fix that" in the reviews so I'd like to turn it over to you all to review the changes I did make, e.g., are the basic demo graphs acceptable?

I did use `ggplot2` so they are a bit sleeker right out of the box than base graphics (added package only to `DESCRIPTION` `Suggests` field). But I did keep them very minimal just like any use of `dplyr` or `tidyr` in the vignettes because I don't want to draw attention away from the tutorial message with too much extraneous code.

Is there anything you all firmly want changed that I missed? Should we continue discussion on certain topics?

Thank you!
