### 2019-10-28 #342 @daranzolin
@leonawicz graphs look good. No further comments or suggestions from me. But as for our previous discussion on cascading selectInputs in Shiny, [here's a good example.](https://gist.github.com/MarkEdmondson1234/9032144)

@melvidoni I updated my original review comment confirming that the package author has responded to my review.
