### 2019-10-28 #342 @ottlngr
@leonawicz Thanks for your detailed answers and sharing your reasoning, especially on the `tidy` argument. I can totally relate to it and it was not meant to be a "change this, fix that" comment, anyway.

No further questions, suggestions or comments.

> - [x] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

@melvidoni I updated my review.
