### 2019-10-28 #342 @leonawicz
Hi all, @ottlngr @daranzolin thanks so much for your help reviewing the package! And thanks @melvidoni for your guidance. I have transferred the repo and made the package changes affected by the transfer. It appears my encrypted API key in `travis.yml` may still work (?) but I'll only know for sure after I get codecov.io integration working again. Please let me know when the admin status is active. Thanks! :)
