### 2019-10-29 #342 @melvidoni
Hello @leonawicz! I already gave you admin rights on your repo, and invited you to an rOpenSci team. Can I publish this in Twitter? If so, please, let me know your handle.
