### 2019-10-29 #342 @leonawicz
Hi @melvidoni I see the team and settings tab now, thanks! Yes, we're all set to move forward. I just pushed a `cran-comments.md` update for the next CRAN release and it looks like Travis-CI and CodeCov are still working as intended after the transfer. My Twitter handle is the same. I plan to mention on Twitter as well. Thanks!
